+++ tests/package-imports.test.ts
@@ -137,7 +137,8 @@ describe("package public imports", () => {
 			  "@tsuuanmi/pi-workflows./tools/workflow-tools",
 			  "@tsuuanmi/pi.",
 			  "@tsuuanmi/pi./api/types",
-			  "@tsuuanmi/pi./config/config",
+			  "@tsuuanmi/pi./loader",
+			  "@tsuuanmi/pi./loader/config",
 			]
 		`);
 	});